[STORM-3685] Move validateCycleFree() to Utils and throw WrappedInvalidTopologyException.
